other_tasks/ai-research-assistant-api/requirements.txt
alembic==1.16.5
fastapi==0.116.1
huggingface-hub==0.35.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-huggingface==0.3.1
langchain-postgres==0.0.15
langgraph==0.6.7
pgvector==0.3.6
psycopg2-binary==2.9.10
PyPDF2==3.0.1
python-dotenv==1.1.1
requests==2.32.5
sentence-transformers==5.1.0
SQLAlchemy==2.0.43
uvicorn==0.35.0
passlib==1.7.4
python-jose==3.5.0
python-multipart==0.0.20
bcrypt>=4.1.1
